docker/src/requirements.txt
Protobuf
grpcio==1.6.3